ref: Extract ReviewFlow into useReviewFlow hook and StarRatingInput component
Replace manual useEffect/useState with SWR for movie fetching, lazy confetti
initialization, and extract star rating widget into a standalone component.
ReviewFlow is now a pure rendering component.
